enhancement: Updater.cs TODO
Updater will check if any files are missing after the initial download, by comparing the list of files that were requested for download with the files that actually exist on the local machine. If there are any missing files, it will add them to the download queue and return false. This will allow the basic downloader to fetch them.